refactor(ui): decompose EntityExtraMeepleCard into entity-specific files
Split the 2052-line monolith into focused modules:
- shared.tsx: ENTITY_COLORS, EntityHeader, EntityTabTrigger, StatCard, loading/error states
- entities/GameExtraMeepleCard.tsx (322 lines + GameAgentCard helper)
- entities/PlayerExtraMeepleCard.tsx (158 lines)
- entities/CollectionExtraMeepleCard.tsx (119 lines)
- entities/AgentExtraMeepleCard.tsx (379 lines + hooks + helpers)
- entities/ChatExtraMeepleCard.tsx (346 lines + ChatBubble helper)
- entities/KbExtraMeepleCard.tsx (488 lines + timeline helpers)

EntityExtraMeepleCard.tsx reduced to barrel re-exports preserving all
existing import paths. No functional changes.
